+++ docs/CHANGELOG.md
@@ -2,9 +2,7 @@
 
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
     
-## [Unreleased]
-
-Making BIDScoin more powerful and generic...
+## [3.6.0] - 2021-05-13
 
 ### Added
 - Support for BIDS v1.6.0 (-> PET)